+++ lemon/__init__.py
@@ -1,0 +1,9 @@
+from .options import ModelAdmin, AppAdmin
+from .options import StackedInline, TabularInline
+from .sites import AdminSite, site
+
+
+__all__ = [
+    'ModelAdmin', 'AppAdmin', 'StackedInline', 'TabularInline', 'AdminSite',
+    'site',
+]