+++ lemon/settings.py
@@ -4,4 +4,4 @@ from django.conf import settings
 CONFIG = {
     'MARKUP_WIDGET': None,
 }
-CONFIG.update(getattr(settings, 'EXTRADMIN_CONFIG', {}))
+CONFIG.update(getattr(settings, 'LEMON_CONFIG', {}))