+++ lemon/templatetags/main_menu.py
@@ -2,7 +2,7 @@ from django.template import Library, Variable
 from django.template import TemplateSyntaxError, VariableDoesNotExist
 from django.template.defaulttags import URLNode
 
-from lemon.extradmin.models import MenuItem
+from ..models import MenuItem
 
 
 register = Library()
@@ -30,7 +30,7 @@ class MainMenuItemURLNode(URLNode):
         return super(MainMenuItemURLNode, self).render(context)
 
 
-@register.inclusion_tag('extradmin/main_menu.html')
+@register.inclusion_tag('lemon/main_menu.html')
 def main_menu():
     queryset = MenuItem.objects.select_related('section', 'content_type')
     queryset = queryset.order_by('section__position', 'position')